requirements.txt
# HTTP request library
requests==2.31.0

# Data analysis
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Visualization
matplotlib==3.7.2

# Environment variables
python-dotenv==1.0.0

# Reddit API wrapper
praw==7.7.0

# Machine learning and forecasting
scikit-learn==1.3.0
statsmodels==0.14.0
prophet==1.1.2

# Pre-commit hooks
pre-commit==3.3.3  # Pre-commit version for managing Git hooks
black==23.3.0  # Python code formatter
flake8==6.1.0  # Python linter
